refactoring the code to use packages
