+++ advanced-go/constants/constants.go
@@ -4,8 +4,8 @@ package constants
 const (
 	PORT_NUMBER = ":9094"
 	// Using actual file path
-	SERVER_CERT = "/media/aniket/9c8cbcfa-7a63-4b30-910a-2586f19beb53/Aniket/Documents/GIT ANIKET/advanced-go/ssl/server.crt"
-	SERVER_KEY  = "/media/aniket/9c8cbcfa-7a63-4b30-910a-2586f19beb53/Aniket/Documents/GIT ANIKET/advanced-go/ssl/server.key"
-	// SERVER_CERT = "./ssl/server.crt"  // this is not working (unable to locate the file)
-	// SERVER_KEY  = "./ssl/server.key"  // this is not working (unable to locate the file)
+	// SERVER_CERT = "/media/aniket/9c8cbcfa-7a63-4b30-910a-2586f19beb53/Aniket/Documents/GIT ANIKET/advanced-go/ssl/server.crt"
+	// SERVER_KEY  = "/media/aniket/9c8cbcfa-7a63-4b30-910a-2586f19beb53/Aniket/Documents/GIT ANIKET/advanced-go/ssl/server.key"
+	SERVER_CERT = "./ssl/server.crt" // this is not working (unable to locate the file)
+	SERVER_KEY  = "./ssl/server.key" // this is not working (unable to locate the file)
 )
