+++ advanced-go/main.go
@@ -5,13 +5,14 @@ import (
 	"net/http"
 
 	"github.com/anik36/Golang/tree/advanced-go/constants"
+	"github.com/anik36/Golang/tree/advanced-go/handlers"
 )
 
 func main() {
 
-	http.HandleFunc("/home", Home)
-	http.HandleFunc("/", Home)
-	http.HandleFunc("/about", About)
+	http.HandleFunc("/home", handlers.Home)
+	http.HandleFunc("/", handlers.Home)
+	http.HandleFunc("/about", handlers.About)
 
 	fmt.Println("Starting application on port", constants.PORT_NUMBER)
 	_ = http.ListenAndServeTLS(constants.PORT_NUMBER, constants.SERVER_CERT, constants.SERVER_KEY, nil)
